backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
python-dotenv==1.0.1
google-generativeai==0.7.2
cachetools==5.3.3
passlib==1.7.4
google-auth==2.30.0
firebase-admin==6.5.0
google-cloud-firestore==2.16.0